moved js files from test-directories
