+++ lib/builders/model-builder.js
@@ -1,6 +1,6 @@
 import { any } from '../util/database';
-import { MappingStrategyFactory } from './factories/mapping-strategy-factory';
-import { ResourceEntityFactory } from './factories/resource-entity-factory';
+import { MappingStrategyFactory } from '../factories/mapping-strategy-factory';
+import { ResourceEntityFactory } from '../factories/resource-entity-factory';
 
 export class ModelBuilder {
