+++ lib/entities/property.js
