+++ lib/entities/reference.js
