+++ lib/entities/resource.js
