+++ lib/factories/mapping-strategy-factory.js
@@ -1,7 +1,7 @@
 
 /** STRATEGIES **/
-import { MapResourceStrategy } from './strategies/map-resource-strategy';
-import { MapPropertyStrategy } from './strategies/map-property-strategy';
+import { MapResourceStrategy } from '../strategies/map-resource-strategy';
+import { MapPropertyStrategy } from '../strategies/map-property-strategy';
 
 export class MappingStrategyFactory {
