+++ lib/factories/object-entity-factory.js
