+++ lib/factories/resource-entity-factory.js
