+++ lib/strategies/map-property-strategy.js
@@ -1,4 +1,4 @@
-import { Property } from '../../entities/property';
+import { Property } from '../entities/property';
 import { MapStrategy } from './map-strategy';
 
 export class MapPropertyStrategy extends MapStrategy{
