+++ lib/strategies/map-resource-strategy.js
