+++ lib/strategies/map-strategy.js
